Rewind log file reading on demand
The file_reader module starts to read files from the end by default. To
start the reading from the beginning, you must rewind this process
manually, which has not yet been implemented.

To do this, now, you may use the new button added to the module's UI to
reset the reading position for the chosen file.